requirements.txt
google-genai>=1.0.0
crawl4ai>=0.4.0
playwright>=1.44.0